:rocket add foundry

+++ src/BouncerProxy.sol
@@ -21,7 +21,6 @@ pragma solidity ^0.8.9;
     uport-identity
 */
 
-
 //use case 1:
 //you deploy the bouncer proxy and use it as a standard identity for your own etherless accounts
 //  (multiple devices you don't want to store eth on or move private keys to will need to be added as Bouncers)
@@ -45,9 +44,8 @@ pragma solidity ^0.8.9;
 //you run your own relayer and pay for all of their transactions, revoking any bad actors if needed
 
 contract BouncerProxy {
-
     //to avoid replay
-    mapping(address => uint) public nonce;
+    mapping(address => uint256) public nonce;
 
     // allow for third party metatx account to make transactions through this
     // contract like an identity but make sure the owner has whitelisted the tx
@@ -55,112 +53,93 @@ contract BouncerProxy {
 
     event UpdateWhitelist(address _account, bool _value);
     // copied from https://github.com/uport-project/uport-identity/blob/develop/contracts/Proxy.sol
-    event Received (address indexed sender, uint value);
+    event Received(address indexed sender, uint256 value);
     // when some frontends see that a tx is made from a bouncerproxy, they may want to parse through these events to find out who the signer was etc
-    event Forwarded (bytes sig, address signer, address destination, uint value, bytes data, bytes32 _hash);
+    event Forwarded(bytes sig, address signer, address destination, uint256 value, bytes data, bytes32 _hash);
 
     //whitelist the deployer so they can whitelist others
-    constructor() 
-    {
+    constructor() {
         whitelist[msg.sender] = true;
     }
 
-    fallback () external payable 
-    {
-        emit Received(msg.sender, msg.value); 
+    fallback() external payable {
+        emit Received(msg.sender, msg.value);
     }
 
-    receive () external payable 
-    {
-        emit Received(msg.sender, msg.value); 
+    receive() external payable {
+        emit Received(msg.sender, msg.value);
     }
 
-    function updateWhitelist(address _account, bool _value) 
-    public 
-    returns(bool) 
-    {
-        require(whitelist[msg.sender],"BouncerProxy::updateWhitelist Account Not Whitelisted");
+    function updateWhitelist(address _account, bool _value) public returns (bool) {
+        require(whitelist[msg.sender], "BouncerProxy::updateWhitelist Account Not Whitelisted");
         whitelist[_account] = _value;
-        emit UpdateWhitelist(_account,_value);
+        emit UpdateWhitelist(_account, _value);
         return true;
     }
 
-
-    function getHash(address signer, address destination, uint value, bytes memory data) 
-    public 
-    view 
-    returns(bytes32)
+    function getHash(address signer, address destination, uint256 value, bytes memory data)
+        public
+        view
+        returns (bytes32)
     {
         return keccak256(abi.encodePacked(address(this), signer, destination, value, data, nonce[signer]));
     }
 
-
     // original forward function copied from https://github.com/uport-project/uport-identity/blob/develop/contracts/Proxy.sol
-    function forward(bytes memory sig, address signer, address destination, uint value, bytes memory data) 
-    public 
-    {
-      //the hash contains all of the information about the meta transaction to be called
-      bytes32 _hash = getHash(signer, destination, value, data);
-      //increment the hash so this tx can't run again
-      nonce[signer]++;
-      //this makes sure signer signed correctly AND signer is a valid bouncer
-      require(signerIsWhitelisted(_hash,sig),"BouncerProxy::forward Signer is not whitelisted");
+    function forward(bytes memory sig, address signer, address destination, uint256 value, bytes memory data) public {
+        //the hash contains all of the information about the meta transaction to be called
+        bytes32 _hash = getHash(signer, destination, value, data);
+        //increment the hash so this tx can't run again
+        nonce[signer]++;
+        //this makes sure signer signed correctly AND signer is a valid bouncer
+        require(signerIsWhitelisted(_hash, sig), "BouncerProxy::forward Signer is not whitelisted");
 
-      //execute the transaction with all the given parameters
-      require(executeCall(gasleft(),destination, value, data));
-      emit Forwarded(sig, signer, destination, value, data, _hash);
+        //execute the transaction with all the given parameters
+        require(executeCall(gasleft(), destination, value, data));
+        emit Forwarded(sig, signer, destination, value, data, _hash);
     }
 
-  // copied from https://github.com/uport-project/uport-identity/blob/develop/contracts/Proxy.sol
-  // which was copied from GnosisSafe
-  // https://github.com/gnosis/gnosis-safe-contracts/blob/master/contracts/GnosisSafe.sol
-    function executeCall(uint gasLimit, address to, uint256 value, bytes memory data) 
-    internal 
-    returns (bool success) 
+    // copied from https://github.com/uport-project/uport-identity/blob/develop/contracts/Proxy.sol
+    // which was copied from GnosisSafe
+    // https://github.com/gnosis/gnosis-safe-contracts/blob/master/contracts/GnosisSafe.sol
+    function executeCall(uint256 gasLimit, address to, uint256 value, bytes memory data)
+        internal
+        returns (bool success)
     {
-    assembly {
-               success := call(gasLimit, to, value, add(data, 0x20), mload(data), 0, 0)
-            }
+        assembly {
+            success := call(gasLimit, to, value, add(data, 0x20), mload(data), 0, 0)
+        }
     }
 
-  //borrowed from OpenZeppelin's ESDA stuff:
-  //https://github.com/OpenZeppelin/openzeppelin-solidity/blob/master/contracts/cryptography/ECDSA.sol
-    function signerIsWhitelisted(bytes32 _hash, bytes memory _signature) 
-    internal 
-    view 
-    returns (bool)
-    {
+    //borrowed from OpenZeppelin's ESDA stuff:
+    //https://github.com/OpenZeppelin/openzeppelin-solidity/blob/master/contracts/cryptography/ECDSA.sol
+    function signerIsWhitelisted(bytes32 _hash, bytes memory _signature) internal view returns (bool) {
         bytes32 r;
         bytes32 s;
         uint8 v;
         // Check the signature length
-        if (_signature.length != 65) 
-        {
-          return false;
+        if (_signature.length != 65) {
+            return false;
         }
         // Divide the signature in r, s and v variables
         // ecrecover takes the signature parameters, and the only way to get them
         // currently is to use assembly.
         // solium-disable-next-line security/no-inline-assembly
         assembly {
-          r := mload(add(_signature, 32))
-          s := mload(add(_signature, 64))
-          v := byte(0, mload(add(_signature, 96)))
+            r := mload(add(_signature, 32))
+            s := mload(add(_signature, 64))
+            v := byte(0, mload(add(_signature, 96)))
         }
         // Version of signature should be 27 or 28, but 0 and 1 are also possible versions
-        if (v < 27) 
-        {
-          v += 27;
+        if (v < 27) {
+            v += 27;
         }
         // If the version is correct return the signer address
-        if (v != 27 && v != 28) 
-        {
-          return false;
+        if (v != 27 && v != 28) {
+            return false;
         } else {
-          // solium-disable-next-line arg-overflow
-          return whitelist[ecrecover(keccak256(
-            abi.encodePacked("\x19Ethereum Signed Message:\n32", _hash)
-          ), v, r, s)];
+            // solium-disable-next-line arg-overflow
+            return whitelist[ecrecover(keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", _hash)), v, r, s)];
         }
     }
 }
